HBSD: FreeBSD-PR: 199518 - fix uninitialized field td_sel of struct thread
From the PR:
--
When thread_alloc() allocates struct thread from thread_zone, the field
td_sel is not initialized.
Later in seltdinit(), if td_sel is not NULL, then this field will not
allocate memory.

While not easy to run into the bug in normal configuration, it is easy
to panic when memguard deliberately overwrites the freed memory with 'M'.
--

FreeBSD-PR: 199518
Submitted-by: luke.tw@gmail.com
Signed-off-by: Oliver Pinter <oliver.pinter@hardenedbsd.org>