HBSD: Backport Hardening Code to 10-stable github-issue: #128
This backports all the added features from the hardened/current/master
branch. Some boot hardening code was commented out and will need to be
fixed up. This does not remove the things we removed in
hardened/current/master, like the a.out image activator (among others).
That will happen soon. This also does not add PIE compilation support,
which will also come in soon.

This commit opens the door towards building OPNSense on top of
HardenedBSD.

Sponsored-by: SoldierX
Signed-Off-by: Shawn Webb <shawn.webb@hardenedbsd.org>