HBSD: hbsd-update: support /boot being a symlink
Both base.txz and kernel.txz contain entries for /boot. When tar tries
to extract items for /boot, it does so by rm'ing /boot beforehand if
/boot is a symlink. This can cause issues where /boot is a symlink
pointed to another mountpoint or ZFS pool (not a typical setup, but
also not infrequent).

As a workaround, untar the /boot portion of {base,kernel}.txz to a
temporary directory and sync that over manually (ironically using
tar). Since this is done within /boot, tar doesn't care that /boot is
a symlink. It's only dealing with the files inside of /boot.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
github-issue:	#180
Sponsored-by:	G2, Inc