HBSD: Fix etcupdate integration
Two problems were fixed: a typo and improper setting of the
needsresolve variable. Piping etcupdate's output to a while loop
causes a subshell (lolwut?!?). Setting needsresolve in the subshell
doesn't update the needsresolve variable within the function. So
pipe etcupdate's output to a temporary file and do a simple for...in
loop.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
github-issue:	#171
MFC-to:		10-STABLE
Sponsored-by:	G2, Inc