HBSD: Don't check for ZFS KLD when non-root.
Commit 37a8b2de291f4cb84b5cac86aab8e5edcaa75176 introduced an
interesting issue with ZFS in which the ZFS binaries failed to work as
non-root due to libzfs attempting to check for the ZFS KLD upon
initialization. Since libzfs opens the ZFS dev entry immediately after
checking the KLD load status, if the KLD's not loaded, the dev entry
doesn't exist and library initialization will fail anyways.

So, only check for the ZFS KLD if libzfs is being utilized by root.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
github-issue:	#135