Revert "HBSD: Default jemalloc's lg_chunk to 16 from 21."
This reverts commit 22ac9855f4b8be780c0b82519523313cb936f551.

Setting to 16 is still too aggressive. Certain autoconf checks end up
spinning in an infinite loop. It seems that setting the lg_chunk
variable within the code has other side-effects. More research is
needed.

If we still want to test this patch out, a better place would be the
hardened/current/unstable branch for now.

Signed-off-by:	Shawn webb <shawn.webb@hardenedbsd.org>
MFC-to:		10-STABLE
github-issue:	#124