HBSD: Enable PIEified base for ARM64.
FreeBSD doesn't support Position-Independent Code in the loader stages
on ARM64 (yet?). Move a couple things around and disable PICification
and PIEification of the loader bits on ARM64. I'd assume that the same
thing would need to happen for ARM, but I'd like that architecture to
die the same death I wish upon i386.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>