HBSD: Generate Integriforce rules for base.
This is half of the equation. This adds optional (default to enabled)
generation of Integriforce rules for base in hbsd-update-build(8).
hbsd-update(8) is the other half of the equation.

This feature allows a user to utilize Integriforce for base by
including the ruleset provided by update archives. Please note that
this feature is more aimed at users of ZFS boot environments. Those
who use ZFS boot environments will find that they can apply
Integriforce rules atomically. They will always stay consistent upon
reboot into a new BE.

hbsd-update(8) will create a new directory, /etc/secadm.d, if it
doesn't exist. The ruleset will be placed as a file underneath that:
/etc/secadm.d/base.integriforce.rules.

The manpage for hbsd-update-build(8) will be updated in a follow-up
commit once this feature is set in stone.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
github-issue:	#189
MFC-to:		10-STABLE