HBSD: Perform DNSSEC validation and distribute update hash.
Since HardenedBSD's infrastructure uses DNSSEC, add a new field to the
DNS TXT record: the hash of the update tarball. This allows
hbsd-update to validate the hash of the tarball prior to extracting
it. Validating DNSSEC provides the ability to securely distribute and
validate the update tarball hash, similar in function to a digital
signature.

DNSSEC validation is enabled by default in hbsd-update, but can be
disabled with hbsd-update.conf(5) or passing the -d option.

Those running their own updating infrastructure will want to enable
DNSSEC in order to fully utilize this feature. With bsdtar and
libarchive having issues [1], this feature is of a critical nature.

[1]: https://gist.github.com/lattera/20e44bfc62a7f53ae7c46774215f0943

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
github-issue:	#208
github-issue:	#209
MFC-to:		10-STABLE
MFC-to:		11-STABLE
(cherry picked from commit 3532a3edd079f1d1b0c0518dafbf7150dfe394c9)