HBSD: fix kyue test case sys/kern/kern_copyin
On HardenedBSD, the last page not always mapped in contrast
to FreeBSD, where the last page always mapped as shared page.

To fix this test, which expects the existence of the last page
just map them in at the test start, and unmap them at the end.

MFS: 11-STABLE
MFS: 12-CURRENT
Signed-off-by: Oliver Pinter <oliver.pinter@hardenedbsd.org>