HBSD: Support extracting hardlink data in libarchive
Introduuce a new per-archive flag in libarchive,
INSECURE_HARDLINK_MODE, to allow extracting hardlink data. This flag
is disabled by default and can be enabled by passing in the -i flag to
bsdtar.

This option should be used with extreme caution as it has the
potential to open the user up to "libarchive vulnerability 4" in the
anonymous vulnerability report at [1].

[1]:
http://hardenedbsd.org/~shawn/vulns/freebsd/2016-07-29/libarchive.txt

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Submitted-by:	@pkubaj (patch modified from submission)
github-issue:	#218
MFC-to:		11-STABLE
MFC-to:		10-STABLE