HBSD: Optionally compile devd as a PIE
On systems where /usr is part of the root filesystem, compiling devd
as a Position-Independent Executable (PIE) is safe. devd depends on
libc++, which gets installed into /usr/lib. If /usr is not part of the
root filesystem, then libc++ may not be available when devd starts up.
In more traditional BSD systems, /usr is not part of the root
filesystem.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
github-issue:	#223
MFC-to:		11-STABLE