HBSD: Introducing SafeStack
Compile world with SafeStack from clang/llvm. Right now, only
applications are compiled with SafeStack, not libraries. clang 3.9.0
does not support compiling shared libraries with SafeStack, though a
patch has been proposed that would enable support.

This is very much still a work-in-progress. More research is needed.

Note that with clang's current implementation of SafeStack, ASLR and W^X
are prerequisites. SafeStack currently requires ASLR and W^X in order to
be effective.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>