HBSD: Teach hbsd-update-build to cross-build
This allows us to support non-amd64 architectures with hbsd-update. Of
main interest is arm64, which can use the same kernel config for all
devices.

While testing this work, I came across the inability to create unsigned
updates due to strong post-build sanity checking. Fix unsigned builds.
Unsigned builds should only be done in the case of testing and should
not be used in production (as was the case here).

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
github-issue:	#224
MFC-to:		10-STABLE
MFC-to:		11-STABLE