HBSD: Provide an option to not require binutils from pkg in hbsd-update-build
clang and lld 4.0.0 can build and link world for ARM64. There's no need
to install the aarch64-binutils from pkg if lld is the default linker.
Note that installing aarch64-binutils is still the default behavior. Set
NEED_CROSS_UTILS to 0 in your hbsd-update-build config file if you can
and want to use lld instead.

I'm not marking this as MFC-able as it is currently not clear whether
clang/lld 4.0.0 will be MFC'd to 11-STABLE by upstream FreeBSD.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	SoldierX