HBSD: Add installation hook scripting
Provide pre-installation and post-installation hooks to be overridden by
a file (script.sh) in the update tarball. This will give the update
builder/distributor the ability to perform fix-ups and/or display extra
messages to the user during the update process.

More hooks will be possible, but this is just the start. Ideally, we'd
have hooks for each step of the update process.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	SoldierX
github-issue:	#251