HBSD: Update passwd files with hbsd-update
Previously, hbsd-update instructed etcupdate to skip the passwd files.
With the import of OpenNTPd in base, this meant that users who used
hbsd-update were left without the _ntp user and group and OpenNTPd would
fail to start.

Instead, allow etcupdate to update the passwd and group files. Rebuild
the database if needed.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	SoldierX
github-issue:	#269
MFC-to:		10-STABLE
MFC-to:		11-STABLE