HBSD: allow to set PaX features as jail parameters
The hardening parameters are configureable since this commit
at jail creation time. The same named jail settings provided
as the sysctls are in the hardening. sysctl leaf.

Example jail.conf sniplet:

  exec.start = "/bin/sh /etc/rc";
  exec.stop = "/bin/sh /etc/rc.shutdown";
  exec.clean;
  mount.devfs;

  path = "/usr/jails/$name";
  host.hostname = "$name";

  hbsdnx {
          hardening.pax.segvguard.status = 3;
          hardening.pax.mprotect.status = 3;
          hardening.pax.pageexec.status = 3;
          hardening.pax.aslr.status = 3;
          persist;
  }

Currently the sysctls are mutable after the jail startup,
but in the future this will change. The same is true for
the nested jails. Now the hardening settings are modifiable
for the child jails.

github-issue: #292
MFC-to: 10-STABLE
MFC-to: 11-STABLE
Signed-off-by: Oliver Pinter <oliver.pinter@hardenedbsd.org>
(cherry picked from commit c8fa0e69ffb064787b08b65ccfa3a41fac548be8)
Signed-off-by: Oliver Pinter <oliver.pinter@hardenedbsd.org>