HBSD: Set llvm-ar, llvm-nm, and llvm-objdump as the default ar, nm, and objdump
With commit a809755341beb611398af2cfe84db6762616cc31, we are ready to
set llvm-ar, llvm-nm, and llvm-objdump as the default ar, nm, and
objdump respectively in base. llvm-ar also becomes ranlib.

Some ports fail to package due to inconsistencies between llvm-ar and
binutils ar. We will work with upstream llvm and the port maintainers to
identify and fix those issues. The number of ports that fail to build
via poudriere is small enough that I feel comfortable setting this while
we work out the kinks.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	SoldierX