HBSD: Build libclang_rt.cfi
In preparation for Cross-DSO CFI, we need to build the Cross-DSO CFI
runtime. We force compilation with LTO since Cross-DSO CFI requires all
to-be-linked objects to be compiled with LTO.

We're still a long ways off from Cross-DSO CFI support in HardenedBSD.
Building libclang_rt.cfi will allow some soak time so that when
Cross-DSO CFI lands, lld will be happy.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
github-issue:	#247