HBSD: Temporarily disable OFED
There's a problem building the ofed utilities right now. I suspect it's
because of LTO being applied to applications and not libraries. The
issue is not manifest on the Cross-DSO CFI branch.

I plan to re-enable OFED when I get more time to diagnose the breakage
and come up with a proper fix.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Sponsored-by:	SoldierX