HBSD: bhyve hardening: Use MAP_GUARD pages as guard pages
The memory space of virtual machines is protected by guard pages.
Instead of relying solely on PROT_NONE mappings, use MAP_GUARD mappings
as the guard pages.

This is the second, more efficient and complete version of the patch.
Thanks to John Baldwin for suggesting I first map the entire range as
MAP_GUARD, then mmap the middle with regular pages.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>